+++ src/elastic-san/azext_elastic_san/aaz/latest/elastic_san/volume_group/_update.py
@@ -34,9 +34,9 @@ class Update(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2024-06-01-preview",
+        "version": "2024-07-01-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.elasticsan/elasticsans/{}/volumegroups/{}", "2024-06-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.elasticsan/elasticsans/{}/volumegroups/{}", "2024-07-01-preview"],
         ]
     }
 
@@ -82,6 +82,28 @@ class Update(AAZCommand):
                 pattern="^[A-Za-z0-9]+((-|_)[a-z0-9A-Z]+)*$",
                 max_length=63,
                 min_length=3,
+            ),
+        )
+
+        # define Arg Group "DeleteRetentionPolicy"
+
+        _args_schema = cls._args_schema
+        _args_schema.delete_retention_policy_state = AAZStrArg(
+            options=["--delete-retention-state", "--delete-retention-policy-state"],
+            arg_group="DeleteRetentionPolicy",
+            help="Manage delete retention policy state",
+            is_preview=True,
+            nullable=True,
+            enum={"Disabled": "Disabled", "Enabled": "Enabled"},
+        )
+        _args_schema.delete_retention_period_days = AAZIntArg(
+            options=["--retention-period", "--delete-retention-period-days"],
+            arg_group="DeleteRetentionPolicy",
+            help="The number of days to retain the resources after deletion.",
+            is_preview=True,
+            nullable=True,
+            fmt=AAZIntArgFormat(
+                minimum=0,
             ),
         )
 
@@ -292,7 +314,7 @@ class Update(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2024-06-01-preview",
+                    "api-version", "2024-07-01-preview",
                     required=True,
                 ),
             }
@@ -395,7 +417,7 @@ class Update(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2024-06-01-preview",
+                    "api-version", "2024-07-01-preview",
                     required=True,
                 ),
             }
@@ -453,7 +475,7 @@ class Update(AAZCommand):
                 value=instance,
                 typ=AAZObjectType
             )
-            _builder.set_prop("identity", AAZObjectType, ".identity")
+            _builder.set_prop("identity", AAZIdentityObjectType, ".identity")
             _builder.set_prop("properties", AAZObjectType, typ_kwargs={"flags": {"client_flatten": True}})
 
             identity = _builder.get(".identity")
@@ -467,11 +489,17 @@ class Update(AAZCommand):
 
             properties = _builder.get(".properties")
             if properties is not None:
+                properties.set_prop("deleteRetentionPolicy", AAZObjectType)
                 properties.set_prop("encryption", AAZStrType, ".encryption")
                 properties.set_prop("encryptionProperties", AAZObjectType, ".encryption_properties")
                 properties.set_prop("enforceDataIntegrityCheckForIscsi", AAZBoolType, ".enforce_data_integrity_check_for_iscsi")
                 properties.set_prop("networkAcls", AAZObjectType, ".network_acls")
                 properties.set_prop("protocolType", AAZStrType, ".protocol_type")
+
+            delete_retention_policy = _builder.get(".properties.deleteRetentionPolicy")
+            if delete_retention_policy is not None:
+                delete_retention_policy.set_prop("policyState", AAZStrType, ".delete_retention_policy_state")
+                delete_retention_policy.set_prop("retentionPeriodDays", AAZIntType, ".delete_retention_period_days")
 
             encryption_properties = _builder.get(".properties.encryptionProperties")
             if encryption_properties is not None:
@@ -578,7 +606,7 @@ class _UpdateHelper:
         volume_group_read.id = AAZStrType(
             flags={"read_only": True},
         )
-        volume_group_read.identity = AAZObjectType()
+        volume_group_read.identity = AAZIdentityObjectType()
         volume_group_read.name = AAZStrType(
             flags={"read_only": True},
         )
@@ -624,6 +652,9 @@ class _UpdateHelper:
         )
 
         properties = _schema_volume_group_read.properties
+        properties.delete_retention_policy = AAZObjectType(
+            serialized_name="deleteRetentionPolicy",
+        )
         properties.encryption = AAZStrType()
         properties.encryption_properties = AAZObjectType(
             serialized_name="encryptionProperties",
@@ -644,6 +675,14 @@ class _UpdateHelper:
         properties.provisioning_state = AAZStrType(
             serialized_name="provisioningState",
             flags={"read_only": True},
+        )
+
+        delete_retention_policy = _schema_volume_group_read.properties.delete_retention_policy
+        delete_retention_policy.policy_state = AAZStrType(
+            serialized_name="policyState",
+        )
+        delete_retention_policy.retention_period_days = AAZIntType(
+            serialized_name="retentionPeriodDays",
         )
 
         encryption_properties = _schema_volume_group_read.properties.encryption_properties