+++ src/containerapp/HISTORY.rst
@@ -5,6 +5,7 @@ Release History
 upcoming
 ++++++
 * 'az containerapp env ingress': New environment level ingress configuration.
+* 'az containerapp auth update': Support authenticating blob storage token store using managed identity with `--blob-container-uri` and `--blob-container-identity`.
 
 1.1.0b4
 ++++++